move Initializer into its own dir
